This is pages 5 and 6 (Login and Register)
The login page contains the request for the user to insert his/her username and password, it also gives the option to to register if not signed up and the option to retrieve forgotten password.

The Register Page form simply asks the user to write his/her first name, enter a  password, as well as confirming the password. There is also a nice little submit button below.